refactor: move router in separate class
